refactor(engine): align REST-SSZ types with validated Reth codecs
