+++ crates/rpc-types-engine/Cargo.toml
@@ -28,7 +28,6 @@ alloy-rlp = { workspace = true, features = ["arrayvec", "derive"] }
 alloy-primitives = { workspace = true, features = ["rlp", "serde"] }
 alloy-consensus = { workspace = true, features = ["serde"] }
 alloy-eips = { workspace = true, features = ["serde"] }
-alloy-rpc-types-debug = { workspace = true, optional = true }
 
 # misc
 derive_more = { workspace = true, features = ["display"] }
@@ -41,7 +40,6 @@ serde = { workspace = true, features = ["derive"], optional = true }
 # ssz
 ethereum_ssz_derive = { workspace = true, optional = true }
 ethereum_ssz = { workspace = true, optional = true }
-ssz_types = { workspace = true, optional = true }
 
 # arbitrary
 arbitrary = { workspace = true, features = ["derive"], optional = true }
@@ -72,24 +70,9 @@ serde = [
 ]
 jwt = ["std", "dep:jsonwebtoken", "dep:rand"]
 jwt-aws-lc-rs = ["jwt", "jsonwebtoken/aws_lc_rs"]
-ssz = [
-    "std",
-    "dep:alloy-rpc-types-debug",
-    "dep:ethereum_ssz",
-    "dep:ethereum_ssz_derive",
-    "dep:ssz_types",
-    "alloy-eips/ssz",
-]
+ssz = ["std", "dep:ethereum_ssz", "dep:ethereum_ssz_derive", "alloy-eips/ssz"]
 kzg = ["alloy-consensus/kzg", "alloy-eips/kzg"]
-arbitrary = [
-	"dep:arbitrary",
-	"std",
-	"alloy-primitives/arbitrary",
-	"alloy-serde/arbitrary",
-	"alloy-consensus/arbitrary",
-	"alloy-eips/arbitrary",
-	"ssz_types?/arbitrary"
-]
+arbitrary = ["dep:arbitrary", "std", "alloy-primitives/arbitrary", "alloy-serde/arbitrary", "alloy-consensus/arbitrary", "alloy-eips/arbitrary"]
 
 [dev-dependencies]
 alloy-primitives = { workspace = true, features = [
